+++ packages/app/src/App.tsx
@@ -4,7 +4,7 @@ import { fetchModulesList } from "./store/modules"
 import { useSafeAppsSDK } from "@gnosis.pm/safe-apps-react-sdk"
 import { useRootDispatch } from "./store"
 import Panel from "./views/Panel"
-import { Views } from "./Views"
+import { Views } from "./View"
 import Header from "./views/Header"
 import { makeStyles } from "@material-ui/core"
 import TransactionBuilder from "./views/TransactionBuilder"
